How to Train a Registry.
