+++ src/main/java/org/spongepowered/common/registry/type/GamemodeRegistryModule.java
@@ -35,12 +35,12 @@ import org.spongepowered.api.entity.living.player.gamemode.GameModes;
 import org.spongepowered.common.registry.AdditionalRegistration;
 import org.spongepowered.common.registry.CatalogRegistryModule;
 import org.spongepowered.common.registry.RegisterCatalog;
-import org.spongepowered.common.registry.Registration;
+import org.spongepowered.common.registry.Registry;
 
 import java.util.Collection;
 import java.util.Optional;
 
-@Registration
+@Registry
 public class GameModeRegistryModule implements CatalogRegistryModule<GameMode> {
 
     @RegisterCatalog(GameModes.class)
