+++ src/main/java/org/spongepowered/common/registry/type/LogAxisRegistryModule.java
@@ -34,12 +34,12 @@ import org.spongepowered.api.data.type.LogAxes;
 import org.spongepowered.api.data.type.LogAxis;
 import org.spongepowered.common.registry.CatalogRegistryModule;
 import org.spongepowered.common.registry.RegisterCatalog;
-import org.spongepowered.common.registry.Registration;
+import org.spongepowered.common.registry.Registry;
 
 import java.util.Collection;
 import java.util.Optional;
 
-@Registration
+@Registry
 public class LogAxisRegistryModule implements CatalogRegistryModule<LogAxis> {
 
     @RegisterCatalog(LogAxes.class)
