+++ src/main/java/org/spongepowered/common/registry/type/RotationRegistryModule.java
@@ -30,14 +30,14 @@ import org.spongepowered.api.util.rotation.Rotation;
 import org.spongepowered.api.util.rotation.Rotations;
 import org.spongepowered.common.registry.CatalogRegistryModule;
 import org.spongepowered.common.registry.RegisterCatalog;
-import org.spongepowered.common.registry.Registration;
+import org.spongepowered.common.registry.Registry;
 import org.spongepowered.common.rotation.SpongeRotation;
 
 import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 
-@Registration
+@Registry
 public class RotationRegistryModule implements CatalogRegistryModule<Rotation> {
 
     @RegisterCatalog(Rotations.class)
